requirements.txt
streamlit
google-generativeai
st-gsheets-connection
pandas
pillow
